builder/backend/requirements.txt
Flask
Flask-Cors
python-dotenv
tframex # Assuming tframex is installable via pip
# Add specific versions if needed, e.g., Flask==2.3.0